Authorization, Gate - 1

+++ app/Http/Controllers/PostController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\StorePost;
 use App\Models\BlogPost;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
@@ -107,6 +108,12 @@ class PostController extends Controller
     public function update(StorePost $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+
+        // if(Gate::denies('update-post', $post)){
+        //     abort(403,"Ban khong the update blog");
+        // }
+        $this->authorize('update-post', $post);
+
         $validated = $request->validated();
         $post->fill($validated);
         $post->save();
@@ -125,6 +132,10 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = BlogPost::findOrFail($id);
+        // if(Gate::denies('delete-post', $post)){
+        //     abort(403,"Ban khong the xoa blog");
+        // }
+        $this->authorize('delete-post', $post);
         $post->delete();
 
         session()->flash('status', 'Blog post was deleted!');
